+++ examples/drgst/dr_gst.py
@@ -1,17 +1,19 @@
 import os
-os.environ['CUDA_VISIBLE_DEVICES'] = '1'
+os.environ['CUDA_VISIBLE_DEVICES'] = '0'
 os.environ['TL_BACKEND'] = 'torch'
+import argparse
+import copy
 import numpy as np
 # tensorlayerx\optimizers\torch_optimizers.py 264行 loss.bakcward() -> loss.backward(retain_graph=True)
 import tensorlayerx as tlx
 import torch
-import tensorflow as tf
 from tensorlayerx.model import TrainOneStep, WithLoss
 from tensorlayerx.losses import mean_squared_error
 from gammagl.datasets import Planetoid
 from gammagl.datasets import Amazon
 from gammagl.datasets import Coauthor, Flickr
-from gammagl.models import GCNModel, GATModel, GraphSAGE_Full_Model, APPNPModel, GINModel, SGCModel, MIXHOPModel, DR_GST
+from gammagl.models import GCNModel, GATModel, GraphSAGE_Full_Model, APPNPModel, GINModel, SGCModel, MIXHOPModel
+from gammagl.transforms import DropEdge
 from gammagl.utils import add_self_loops, mask_to_index
 from gammagl.data import Dataset
 import random
@@ -70,21 +72,21 @@ class DR_GST():
         return acc_test, loss_test
 
     @staticmethod
-    def generate_mask(dataset, labels, labelrate, os_path=None):
-        datalength = labels.size()[0]
-        train_mask, val_mask, test_mask = torch.full((1, datalength), fill_value=False, dtype=bool), torch.full(
-            (1, datalength), fill_value=False, dtype=bool) \
-            , torch.full((1, datalength), fill_value=False, dtype=bool)
-        name = ('train', 'val', 'test')
+    def generate_mask(labels):
+        
+        length = tlx.get_tensor_shape(labels)[0]
+        train_mask = tlx.zeros(shape = (1, length), dtype = tlx.bool)
+        val_mask = tlx.zeros(shape = (1, length), dtype = tlx.bool)
+        test_mask = tlx.zeros(shape = (1, length), dtype = tlx.bool)
 
-        indices = torch.randperm(datalength)
+        indices = np.random.permutation(length)
+        indices = tlx.convert_to_tensor(indices)
+        # indices = torch.randperm(datalength)
 
         train_ratio = 0.8
         val_ratio = 0.1
-        test_ratio = 0.1
-        train_size = int(train_ratio * datalength)
-        val_size = int(val_ratio * datalength)
-        test_size = datalength - train_size - val_size
+        train_size = int(train_ratio * length)
+        val_size = int(val_ratio * length)
 
         train_mask[0][indices[:train_size]] = True
         val_mask[0][indices[train_size:train_size + val_size]] = True
@@ -96,10 +98,11 @@ class DR_GST():
 
     @staticmethod
     def update_T(T, output, idx_train, labels):
+        
         if tlx.BACKEND == 'torch':
             output = torch.softmax(output, dim=1)
             T.requires_grad = True
-            optimizer = torch.optim.Adam([T], lr=0.01, weight_decay=5e-4)
+            optimizer = torch.optim.Adam([T], lr=0.001, weight_decay=5e-4)
             mse_criterion = torch.nn.MSELoss()
             index = torch.where(idx_train)[0]
             nclass = labels.max().item() + 1
@@ -126,8 +129,8 @@ class DR_GST():
         data = dataset.data
         train_mask = data.train_mask
         nclass = dataset.num_classes
-        if torch.is_tensor(train_mask) == False:
-            train_mask = torch.tensor(train_mask)
+        if not tlx.is_tensor(train_mask) == False:
+            train_mask = tlx.convert_to_tensor(train_mask)
         train_index = mask_to_index(train_mask)
         train_mask = train_mask.clone()
         train_mask[:] = False
@@ -150,13 +153,13 @@ class DR_GST():
         train_idx = idx_train
 
         test_mask = data.test_mask
-        if torch.is_tensor(test_mask) == False:
-            test_mask = torch.tensor(test_mask)
+        if not tlx.is_tensor(test_mask):
+            test_mask = tlx.convert_to_tensor(test_mask)
         test_idx = mask_to_index(test_mask)
 
         val_mask = data.val_mask
-        if torch.is_tensor(val_mask) == False:
-            val_mask = torch.tensor(val_mask)
+        if not tlx.is_tensor(val_mask) == False:
+            val_mask = tlx.convert_to_tensor(val_mask)
 
         val_idx = mask_to_index(val_mask)
 
@@ -199,11 +202,13 @@ class DR_GST():
                   + "  val loss: {:.4f}".format(loss_val) \
                   + "  val acc: {:.4f}".format(acc_val))
 
-            model_path = './save_model/%s-%s-%d-%f-%f-%f-%s.pth' % (
-                args.model, args.dataset, args.labelrate, args.threshold, args.beta, args.droprate, args.drop_method)
+            # model_path = './save_model/%s-%s-%d-%f-%f-%f-%s.pth' % (
+            #     args.model, args.dataset, args.labelrate, args.threshold, args.beta, args.droprate, args.drop_method)
 
             if loss_val < best:
-                tlx.files.save_npz(net.all_weights, name=model_path + net.name + ".npz")
+                print("-------------------------")
+                print(args.best_model_path + net.name + ".npz")
+                tlx.files.save_npz(net.all_weights, name=args.best_model_path + net.name + ".npz")
                 best = loss_val
                 bad_counter = 0
                 best_output = logits
@@ -267,16 +272,16 @@ class DR_GST():
         tlx.files.load_and_assign_npz(name=model_path + net.name + ".npz", network=net)
         out_list = []
         data = dataset.data
-        with torch.no_grad():
-            for _ in range(f_pass):
-                output = net(data.x, data.edge_index, None, data.num_nodes)
-                output = tlx.softmax(output, axis=1)
-                out_list.append(output)
-            out_list = tlx.stack(out_list)
-            out_mean = tlx.reduce_mean(out_list, axis=0)
-            entropy = tlx.reduce_sum(tlx.reduce_mean(out_list * tlx.log(out_list), axis=0), axis=1)
-            Eentropy = tlx.reduce_sum(out_mean * tlx.log(out_mean), axis=1)
-            bald = entropy - Eentropy
+        # with torch.no_grad():
+        for _ in range(f_pass):
+            output = net(data.x, data.edge_index, None, data.num_nodes)
+            output = tlx.softmax(output, axis=1)
+            out_list.append(output)
+        out_list = tlx.stack(out_list)
+        out_mean = tlx.reduce_mean(out_list, axis=0)
+        entropy = tlx.reduce_sum(tlx.reduce_mean(out_list * tlx.log(out_list), axis=0), axis=1)
+        Eentropy = tlx.reduce_sum(out_mean * tlx.log(out_mean), axis=1)
+        bald = entropy - Eentropy
         return bald.detach()
 
 
@@ -363,3 +368,98 @@ class DR_GST():
                                 name="MixHop")
 
         return model
+
+
+def main(args):
+    # model_path = './save_model/%s-%s-%d-%f-%f-%f-%s.pth' % (
+    #     args.model, args.dataset, args.labelrate, args.threshold, args.beta, args.droprate, args.drop_method)
+
+    dataset_path = args.dataset_path
+    dataset_name = args.dataset
+    dataset = DR_GST.load_dataset(dataset_path, dataset_name)
+
+    data = dataset.data
+    labels = data.y
+    num_node = data.num_nodes
+    num_class = dataset.num_classes
+
+    # seed = np.random.randint(1000)
+    # torch.manual_seed(seed)
+    # torch.cuda.manual_seed(seed)
+
+    if dataset_name in {'Cora', 'CiteSeer', 'PubMed', 'CoraFull'}:
+        idx_train = data.train_mask
+        idx_val = data.val_mask
+        idx_test = data.test_mask
+    else:
+        idx_train, idx_val, idx_test = DR_GST.generate_mask(labels=labels)
+
+    if args.drop_method == 'dropedge':
+        drop_edge_data = DropEdge(p=0.1)(data)
+
+    idx_train = DR_GST.generate_trainmask(dataset, labels, args.labelrate)
+
+    idx_train_ag = mask_to_index(idx_train)
+    idx_val_ag = mask_to_index(idx_val)
+    pseudo_labels = copy.deepcopy(labels)
+    bald = tlx.ones((num_node,))
+    T = torch.nn.Parameter(torch.eye(num_class, num_class))
+    T.requires_grad = False
+
+    for s in range(args.stage):
+        best_output = DR_GST.train(dataset, args, bald, T, idx_val_ag, idx_train_ag, pseudo_labels)
+        T = DR_GST.update_T(T.detach(), best_output, idx_train, data.y.data)
+        idx_unlabeled = ~(idx_train | idx_test | idx_val)
+
+        if args.drop_method == 'dropout':
+            bald = DR_GST.uncertainty_dropout(args.best_model_path, args, dataset)
+        elif args.drop_method == 'dropedge':
+            bald = DR_GST.uncertainty_dropedge(drop_edge_data, args.best_model_path, args, dataset)
+
+        net = DR_GST.get_model(args, dataset)
+        tlx.files.load_and_assign_npz(name=args.best_model_path + net.name + ".npz", network=net)
+        if tlx.BACKEND == 'torch':
+            net.to(data['x'].device)
+        net.set_eval()
+
+        best_output = net(data.x, data.edge_index, None, data.num_nodes)
+        idx_train_ag, pseudo_labels, idx_pseudo = DR_GST.regenerate_pseudo_label(best_output, labels, idx_train,
+                                                                                 idx_unlabeled, args.threshold)
+        acc_test, loss_test = DR_GST.test(best_output, data.y, idx_test, num_class)
+
+    return
+
+
+if __name__ == '__main__':
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--model', type=str, default='GCN')
+    parser.add_argument('--dataset', type=str, default="Cora",
+                        help='dataset for training')
+    parser.add_argument("--dataset_path", type=str, default=r'../',
+                        help="path to save dataset")
+    parser.add_argument('--labelrate', type=int, default=20)
+    parser.add_argument('--weight_decay', type=float, default=5e-4,
+                        help='Weight decay (L2 loss on parameters).')
+    parser.add_argument('--epochs', type=int, default=2000,
+                        help='Number of epochs to train.')
+    parser.add_argument('--stage', type=int, default=2)
+    parser.add_argument('--threshold', type=float, default=0.53)
+    parser.add_argument('--beta', type=float, default=1 / 3,
+                        help='coefficient for weighted CE loss')
+    parser.add_argument('--drop_method', type=str, default='dropout')
+    parser.add_argument('--lr', type=float, default=0.01,
+                        help='Initial learning rate.')
+    parser.add_argument('--droprate', type=float, default=0.5,
+                        help='Droprate for MC-Dropout')
+    parser.add_argument('--hidden_dim', type=int, default=64,
+                        help='Number of hidden units.')
+    parser.add_argument('--dropout', type=float, default=0.5,
+                        help='Dropout rate (1 - keep probability).')
+    parser.add_argument('--patience', type=int, default=100)
+    parser.add_argument("--self_loops", type=int, default=1, help="number of graph self-loop")
+    parser.add_argument("--num_layers", type=int, default=2, help="number of layers")
+    parser.add_argument("--norm", type=str, default='both', help="how to apply the normalizer.")
+    parser.add_argument("--best_model_path", type=str, default=r'./', help="path to save best model")
+
+    args = parser.parse_args()
+    main(args)
